+++ Zipangu/NetUnitTest/NetUnitTest.csproj
@@ -54,6 +54,7 @@
     <Compile Include="KanaConversionTest.cs" />
     <Compile Include="AsciiConversionTest.cs" />
     <Compile Include="TestHelper.cs" />
+    <Compile Include="VBStringsTest.cs" />
     <Compile Include="VBStringsHelper.cs" />
   </ItemGroup>
   <ItemGroup>
